requirements.txt
tensorflow==2.19.0
tensorflow-hub>=0.16.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.3.0
Pillow>=9.0.0
jupyter>=1.0.0
ipython>=8.0.0
streamlit>=1.28.0
seaborn==0.13.2
